refactor(stdlib): change dir structure
Change directory structure in order to make easier module inclusion
based on enabled features.
The new `doc` module won't be used with the no-std feature that is going
to be introduced in the next commit.
