+++ language/benchmarks/src/move_vm.rs
@@ -35,12 +35,12 @@ pub fn bench<M: Measurement + 'static>(c: &mut Criterion<M>, fun: &str) {
 
 // Compile `bench.move` and its dependencies
 fn compile_modules() -> Vec<CompiledModule> {
-    let mut src_files = move_stdlib::move_stdlib_files();
+    let mut src_files = move_stdlib::doc::move_stdlib_files();
     src_files.push(MOVE_BENCH_SRC_PATH.to_str().unwrap().to_owned());
     let (_files, compiled_units) = Compiler::from_files(
         src_files,
         vec![],
-        move_stdlib::move_stdlib_named_addresses(),
+        move_stdlib::doc::move_stdlib_named_addresses(),
     )
     .build_and_report()
     .expect("Error compiling...");
