+++ language/move-stdlib/src/tests.rs
@@ -11,9 +11,9 @@ use walkdir::{DirEntry, WalkDir};
 fn check_that_docs_are_updated() {
     let temp_dir = tempdir().unwrap();
 
-    crate::build_stdlib_doc(&temp_dir.path().to_string_lossy());
+    crate::doc::build_stdlib_doc(&temp_dir.path().to_string_lossy());
 
-    let res = check_dirs_not_diff(&temp_dir, &crate::move_stdlib_docs_full_path());
+    let res = check_dirs_not_diff(&temp_dir, &crate::doc::move_stdlib_docs_full_path());
     assert!(
         res.is_ok(),
         "Generated docs differ from the ones checked in. {}",
@@ -27,12 +27,12 @@ fn check_that_docs_are_updated() {
 fn check_that_the_errmap_is_updated() {
     let temp_file = tempfile::NamedTempFile::new().unwrap();
 
-    crate::build_error_code_map(&temp_file.path().to_string_lossy());
+    crate::doc::build_error_code_map(&temp_file.path().to_string_lossy());
 
     assert!(
         file_diff::diff(
             &temp_file.path().to_string_lossy(),
-            &crate::move_stdlib_errmap_full_path()
+            &crate::doc::move_stdlib_errmap_full_path()
         ),
         "Generated errmap differ from the one checked in"
     );
