+++ language/move-stdlib/tests/move_unit_test.rs
@@ -6,7 +6,7 @@ use move_cli::base::test::{run_move_unit_tests, UnitTestResult};
 use move_core_types::account_address::AccountAddress;
 use move_stdlib::{
     natives::{all_natives, nursery_natives, GasParameters, NurseryGasParameters},
-    path_in_crate,
+    doc::path_in_crate,
 };
 use move_unit_test::UnitTestingConfig;
 use tempfile::tempdir;
