+++ language/move-vm/integration-tests/src/compiler.rs
@@ -20,7 +20,7 @@ pub fn compile_units(s: &str) -> Result<Vec<AnnotatedCompiledUnit>> {
     let (_, units) = MoveCompiler::from_files(
         vec![file_path.to_str().unwrap().to_string()],
         vec![],
-        move_stdlib::move_stdlib_named_addresses(),
+        move_stdlib::doc::move_stdlib_named_addresses(),
     )
     .build_and_report()?;
