+++ language/move-vm/runtime/Cargo.toml
@@ -26,7 +26,7 @@ move-vm-types = { path = "../types", default-features = false }
 move-binary-format = { path = "../../move-binary-format", default-features = false }
 
 [features]
-default = ["std", "once_cell/std"]
+default = ["std"]
 fuzzing = ["move-vm-types/fuzzing"]
 failpoints = ["fail/failpoints"]
 # Enable tracing and debugging also for release builds. By default, it is only enabled for debug builds.
@@ -41,4 +41,5 @@ std = [
     "move-core-types/std",
     "move-vm-types/std",
     "move-binary-format/std",
+    "once_cell/std"
 ]
