+++ language/testing-infra/test-generation/src/lib.rs
@@ -56,9 +56,9 @@ fn run_verifier(module: CompiledModule) -> Result<CompiledModule, String> {
 static STORAGE_WITH_MOVE_STDLIB: Lazy<InMemoryStorage> = Lazy::new(|| {
     let mut storage = InMemoryStorage::new();
     let (_, compiled_units) = Compiler::from_files(
-        move_stdlib::move_stdlib_files(),
+        move_stdlib::doc::move_stdlib_files(),
         vec![],
-        move_stdlib::move_stdlib_named_addresses(),
+        move_stdlib::doc::move_stdlib_named_addresses(),
     )
     .build_and_report()
     .unwrap();
