+++ language/testing-infra/transactional-test-runner/src/vm_test_harness.rs
@@ -30,7 +30,7 @@ use move_core_types::{
     value::MoveValue,
 };
 use move_resource_viewer::MoveValueAnnotator;
-use move_stdlib::move_stdlib_named_addresses;
+use move_stdlib::doc::move_stdlib_named_addresses;
 use move_symbol_pool::Symbol;
 use move_vm_runtime::{
     config::VMConfig,
@@ -375,8 +375,8 @@ static PRECOMPILED_MOVE_STDLIB: Lazy<FullyCompiledProgram> = Lazy::new(|| {
     let program_res = move_compiler::construct_pre_compiled_lib(
         vec![PackagePaths {
             name: None,
-            paths: move_stdlib::move_stdlib_files(),
-            named_address_map: move_stdlib::move_stdlib_named_addresses(),
+            paths: move_stdlib::doc::move_stdlib_files(),
+            named_address_map: move_stdlib::doc::move_stdlib_named_addresses(),
         }],
         None,
         move_compiler::Flags::empty(),
@@ -393,9 +393,9 @@ static PRECOMPILED_MOVE_STDLIB: Lazy<FullyCompiledProgram> = Lazy::new(|| {
 
 static MOVE_STDLIB_COMPILED: Lazy<Vec<CompiledModule>> = Lazy::new(|| {
     let (files, units_res) = move_compiler::Compiler::from_files(
-        move_stdlib::move_stdlib_files(),
+        move_stdlib::doc::move_stdlib_files(),
         vec![],
-        move_stdlib::move_stdlib_named_addresses(),
+        move_stdlib::doc::move_stdlib_named_addresses(),
     )
     .build()
     .unwrap();
