+++ language/tools/move-cli/src/main.rs
@@ -7,7 +7,7 @@ use move_core_types::{account_address::AccountAddress, errmap::ErrorMapping};
 use move_stdlib::natives::{all_natives, nursery_natives, GasParameters, NurseryGasParameters};
 
 fn main() -> Result<()> {
-    let error_descriptions: ErrorMapping = bcs::from_bytes(move_stdlib::error_descriptions())?;
+    let error_descriptions: ErrorMapping = bcs::from_bytes(move_stdlib::doc::error_descriptions())?;
     let cost_table = &move_vm_test_utils::gas_schedule::INITIAL_COST_SCHEDULE;
     let addr = AccountAddress::from_hex_literal("0x1").unwrap();
     let natives = all_natives(addr, GasParameters::zeros())
