+++ language/tools/move-unit-test/tests/move_unit_test_testsuite.rs
@@ -84,8 +84,8 @@ fn run_test_impl(path: &Path) -> anyhow::Result<()> {
         num_threads: 1,
         gas_limit: Some(1000),
         source_files,
-        dep_files: move_stdlib::move_stdlib_files(),
-        named_address_values: move_stdlib::move_stdlib_named_addresses()
+        dep_files: move_stdlib::doc::move_stdlib_files(),
+        named_address_values: move_stdlib::doc::move_stdlib_named_addresses()
             .into_iter()
             .collect(),
         report_writeset: true,
